requirements.txt
fastapi
uvicorn
motor
python-dotenv
openai
pydantic